+++ packages/super-editor/src/components/SuperEditor.vue
@@ -5,7 +5,7 @@ import { ref, onMounted, onBeforeUnmount, computed, shallowRef } from 'vue';
 import { Editor } from '@vue-3/index.js';
 import { getStarterExtensions } from '@extensions/index.js';
 import { observeDomChanges } from './pagination-helpers.js';
-import { TextSelection } from 'prosemirror-state';
+import { onMarginClickCursorChange } from './cursor-helpers.js';
 
 const emit = defineEmits(['editor-ready', 'editor-click', 'editor-keydown', 'comments-loaded', 'selection-update']);
 
@@ -152,45 +152,7 @@ onMounted(() => {
 const handleMarginClick = (event) => {
   if (event.target.classList.contains('ProseMirror')) return;
 
-  const y = event.clientY;
-  const x = event.clientX;
-  const { view } = editor.value;
-  const editorRect = view.dom.getBoundingClientRect();
-  
-  let coords = {
-    left: 0,
-    top: y,
-  };
-  let isRightMargin = false;
-  
-  if (x > editorRect.right) {
-    coords.left = editorRect.left + editorRect.width - 1;
-    isRightMargin = true;
-  } else if (x < editorRect.left) {
-    coords.left = editorRect.left;
-  }
-  
-  const pos = view.posAtCoords(coords)?.pos;
-  if (pos) {
-    let cursorPos = pos;
-
-    if (isRightMargin) {
-      const $pos = view.state.doc.resolve(pos);
-      const charOffset = $pos.textOffset;
-      
-      const node = view.state.doc.nodeAt(pos);
-      const text = node?.text;
-      const charAtPos = text?.charAt(charOffset);
-      
-      cursorPos = node?.isText && charAtPos !== ' ' ? pos - 1 : pos;
-    }
-    
-    const transaction = view.state.tr.setSelection(
-      TextSelection.create(view.state.doc, cursorPos)
-    );
-    view.dispatch(transaction);
-    view.focus();
-  }
+  onMarginClickCursorChange(event, editor.value);
 };
 
 onBeforeUnmount(() => {
